control/controlclient, wgengine/filter: extract parsePacketFilter to new constructor in wgengine/filter
